Make PDFs of potential supp figures
- Addresses Dan's email on 2023-03-27

Dan: "Massive outcome, we are delighted. We did get one new comment from reviewer 5. While I think we had a pretty good look at this and didn’t see anything particularly noteworthy, is it worth generating a supp fig showing the 5 clusters in the stage 3’s and listing the top 10 genes for each?"
  Reviewer: "ScRNAseq data has not been sufficiently addressed. Even if the authors can not discriminate between stage 1 and stage 2 in the unbiased clustering of the scRNAseq data, clusters found within the UMAP region labelled as stage 3 should be shown and discussed (shown in their Github notebooks, where 5 clusters are analysed). These clusters may represent transient states, with specific markers, and authors could try to link them to the pseudotime analysis."